Refactor loader package into dataset/model modules
Move and rename loader modules into datasetLoaders/ and modelLoaders/ packages and update import paths across the codebase. Files moved/renamed include datasetLoader -> datasetLoaders/loader, datasetGhost -> datasetLoaders/ghost, modelLoader -> modelLoaders/loader, modelGhost -> modelLoaders/ghost, zeroModel -> modelLoaders/zeroModel, and related references in UI/Plots.py, client/environment.py, and several modules (MACE, Nequip, SchNet, SpookyNet, aseDataset, sGDML). This clarifies separation between dataset and model loader functionality and fixes import paths accordingly.
